Les exeptions sont gérés en RunTimeExeptions

+++ src/main/java/com/example/prestibanquecbhh/services/imp/ClientServicesImp.java
@@ -2,6 +2,7 @@ package com.example.prestibanquecbhh.services.imp;
 
 import com.example.prestibanquecbhh.dtos.ClientDto;
 import com.example.prestibanquecbhh.entities.Client;
+import com.example.prestibanquecbhh.entities.Conseiller;
 import com.example.prestibanquecbhh.repositories.ClientRepository;
 import com.example.prestibanquecbhh.repositories.ConseillerRepository;
 import com.example.prestibanquecbhh.services.ClientServices;
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ClientServicesImp implements ClientServices {
@@ -20,7 +22,11 @@ public class ClientServicesImp implements ClientServices {
 
     @Override
     public List<Client> getAllClient() {
-        return clientRepository.findAll();
+        List<Client> clients = clientRepository.findAll();
+        if (clients.isEmpty()) {
+            throw new RuntimeException("Tu es une mauvaise entreprise ta aucun client ! Bouh");
+        }
+        return clients;
     }
 
     @Override
@@ -32,25 +38,38 @@ public class ClientServicesImp implements ClientServices {
         newClient.setAddress(client.getAddress());
         newClient.setZipcode(client.getZipcode());
         newClient.setPhonenumber(client.getPhonenumber());
+
+        if (conseillerRepository.findById(client.getId_conseiller()).isEmpty()) {
+            throw new RuntimeException("Conseiller avec l'ID correspondant non trouvé.");
+        }
+
         newClient.setConseiller(conseillerRepository.findById(client.getId_conseiller()).get());
+
         return clientRepository.saveAndFlush(newClient);
     }
-
     @Override
     public void deleteClient(Integer id) {
-        clientRepository.deleteById(id);
+        if (clientRepository.existsById(id)) {
+            clientRepository.deleteById(id);
+        } else throw new RuntimeException("Le client n'existe pas");
     }
 
     @Override
     public Client updateClient(Integer id, ClientDto client) {
-        Client existingClient = clientRepository.findById(id).get();
-        existingClient.setLastname(client.getLastname());
-        existingClient.setFirstname(client.getFirstname());
-        existingClient.setCity(client.getCity());
-        existingClient.setAddress(client.getAddress());
-        existingClient.setPhonenumber(client.getPhonenumber());
-        existingClient.setZipcode(client.getZipcode());
-        existingClient.setConseiller(conseillerRepository.findById(client.getId_conseiller()).get());
-        return clientRepository.save(existingClient);
+        if (clientRepository.existsById(id)){
+            Client existingClient = clientRepository.findById(id).get();
+            existingClient.setLastname(client.getLastname());
+            existingClient.setFirstname(client.getFirstname());
+            existingClient.setCity(client.getCity());
+            existingClient.setAddress(client.getAddress());
+            existingClient.setPhonenumber(client.getPhonenumber());
+            existingClient.setZipcode(client.getZipcode());
+            if (conseillerRepository.existsById(client.getId_conseiller())){
+                existingClient.setConseiller(conseillerRepository.findById(client.getId_conseiller()).get());
+            } else throw new RuntimeException("Wesh, tu veux changer de conseiller, mais prends en un qui existe !");
+
+            return clientRepository.save(existingClient);
+        } else throw new RuntimeException("Wesh, tu veux modifier un client qui n'existe pas, t'es bizarre toi !");
+
     }
 }
